+++ front/src/recoil/holdStockList/index.ts
@@ -1,6 +1,1 @@
-import atom from './atom';
-import type { IHoldStockItem, IHoldStockListItem } from './atom';
-
-export { IHoldStockItem, IHoldStockListItem };
-
-export default atom;
+export { default } from './atom';
